# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chợ Tốt Clone API",
    "version": "1.0.0",
    "description": "Marketplace/Classifieds REST API (code-first OpenAPI từ Zod).\n\n**Chưa triển khai — mọi method trả HTTP 501:**\n- `/uploads — upload ảnh lên S3`\n- `/search — full-text search`\n- `/reviews — đánh giá người bán`"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "CreateCategory": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 60,
            "example": "Sách vở"
          },
          "slug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 60,
            "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
          },
          "icon": {
            "type": "string",
            "maxLength": 8,
            "example": "📚"
          },
          "order": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{24}$"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "icon",
          "order",
          "isActive"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/categories": {
      "get": {
        "operationId": "categoryList",
        "tags": [
          "Category"
        ],
        "summary": "Danh sách danh mục",
        "description": "Từ điển dùng chung toàn hệ thống, không thuộc organization nào. Mặc định chỉ trả danh mục đang bật — dùng `includeInactive=true` khi cần thấy cả danh mục đã ngừng sử dụng.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "includeInactive",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách danh mục",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Query không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createCategory",
        "tags": [
          "Category"
        ],
        "summary": "Tạo danh mục dùng chung toàn hệ thống (chỉ master)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đã tạo danh mục",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Category"
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Cần quyền master",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Slug danh mục đã tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}